[integration] Refactoring
